Removed algorithm.py
- This decouples the structure of ATM and AAM fitters
